Support regular expressions in expected test output
There are cases where the expected output is not static, like the temporary
directories in g-ir-scanner or the home directory in
tests/release-rpm/expected. If an expected output line starts with ^, treat it
as a regular expression instead of a static string, and ensure the whole line
matches (it implicitly appends '$'). ^ is not a valid (or at least reasonable)
program name, so this should not conflict.

Fix tests/release-rpm/expected to not hardcode the directory name.